requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
python-dotenv==1.0.0
bcrypt==5.0.0